Moving IMatrixMul34 to common implementation
